scripts/populate-db: extract out transformer funcs to transformers.py
